refactor: 상금과 관련된 클래스의 위치를 패키지에 위치하도록 변경
