+++ src/main/java/lotto/domain/LottoManager.java
@@ -1,6 +1,11 @@
 package lotto.domain;
 
 import java.util.List;
+import lotto.domain.lotto.Lotto;
+import lotto.domain.lotto.Lottos;
+import lotto.domain.lotto.WinningLotto;
+import lotto.domain.prize.Prize;
+import lotto.domain.prize.WinningDetails;
 
 public class LottoManager {
     private WinningDetails winningDetails;
