+++ src/main/java/lotto/domain/lotto/BonusNumber.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.domain.lotto;
 
 import java.util.List;
 import lotto.util.LottoConstants;
