+++ src/main/java/lotto/domain/lotto/WinningLotto.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.domain.lotto;
 
 import java.util.HashSet;
 import java.util.List;
