+++ src/main/java/lotto/domain/prize/WinningDetails.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.domain.prize;
 
 import java.util.Arrays;
 import java.util.EnumMap;
