+++ src/test/java/lotto/domain/BonusNumberTest.java
@@ -1,11 +1,10 @@
 package lotto.domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.jupiter.api.Assertions.*;
 
 import java.util.Arrays;
 import java.util.List;
-import org.assertj.core.api.Assertions;
+import lotto.domain.lotto.BonusNumber;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
